Refactor code and make actual plots in paper
